refactor(evaluatorq-py): move build_simulation_client to openresponses/client.py (RES-897)
